+++ ChangeLog.md
@@ -6,3 +6,5 @@
 - Added screen break points checker
 - Added universal style getter
 - Added browser Resize Property handler
+- Added child element index getter
+- Added Custom form component builder